1.0.3 commit .

+++ soul-admin/src/main/java/org/dromara/soul/admin/service/impl/PluginServiceImpl.java
@@ -48,6 +48,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.sql.Timestamp;
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
@@ -118,7 +119,10 @@ public class PluginServiceImpl implements PluginService {
                 return AdminConstants.PLUGIN_NAME_IS_EXIST;
             }
         } else {
-            if (exist.getId().equals(pluginDTO.getId())) {
+            if (Objects.isNull(exist)) {
+                return AdminConstants.PLUGIN_NAME_NOT_EXIST;
+            }
+            if (!exist.getId().equals(pluginDTO.getId())) {
                 return AdminConstants.PLUGIN_NAME_IS_EXIST;
             }
         }
@@ -133,15 +137,17 @@ public class PluginServiceImpl implements PluginService {
      */
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public int delete(final List<String> ids) {
-        int pluginCount = 0;
+    public String delete(final List<String> ids) {
         for (String id : ids) {
             PluginDO pluginDO = pluginMapper.selectById(id);
+            if (Objects.isNull(pluginDO)) {
+                return AdminConstants.SYS_PLUGIN_ID_NOT_EXIST;
+            }
             // if sys plugin not delete
             if (pluginDO.getRole().equals(PluginRoleEnum.SYS.getCode())) {
-                return pluginCount;
+                return AdminConstants.SYS_PLUGIN_NOT_DELETE;
             }
-            pluginCount += pluginMapper.delete(id);
+            pluginMapper.delete(id);
             String pluginPath = ZkPathConstants.buildPluginPath(pluginDO.getName());
             if (zkClient.exists(pluginPath)) {
                 zkClient.delete(pluginPath);
@@ -155,8 +161,32 @@ public class PluginServiceImpl implements PluginService {
                 zkClient.delete(ruleParentPath);
             }
         }
-        return pluginCount;
+        return "";
     }
+
+    @Override
+    public String enabled(List<String> ids, Boolean enabled) {
+        for (String id : ids) {
+            PluginDO pluginDO = pluginMapper.selectById(id);
+            if (Objects.isNull(pluginDO)) {
+                return AdminConstants.SYS_PLUGIN_ID_NOT_EXIST;
+            }
+            pluginDO.setDateUpdated(new Timestamp(System.currentTimeMillis()));
+            pluginDO.setEnabled(enabled);
+
+            pluginMapper.updateEnable(pluginDO);
+
+            String pluginPath = ZkPathConstants.buildPluginPath(pluginDO.getName());
+            if (!zkClient.exists(pluginPath)) {
+                zkClient.createPersistent(pluginPath, true);
+            }
+            zkClient.writeData(pluginPath, new PluginZkDTO(pluginDO.getId(),
+                    pluginDO.getName(), pluginDO.getRole(), pluginDO.getEnabled()));
+
+        }
+        return "";
+    }
+
 
     /**
      * find plugin by id.
@@ -212,6 +242,10 @@ public class PluginServiceImpl implements PluginService {
     public int syncPluginAll() {
         List<PluginDO> pluginDOs = pluginMapper.selectByQuery(new PluginQuery());
         if (CollectionUtils.isNotEmpty(pluginDOs)) {
+            String pluginPath = ZkPathConstants.buildPluginParentPath();
+            if (!zkClient.exists(pluginPath)) {
+                zkClient.createPersistent(pluginPath, true);
+            }
             List<String> pluginZKs = zkClient.getChildren(ZkPathConstants.buildPluginParentPath());
             pluginDOs.forEach(pluginDO -> {
                 if (CollectionUtils.isNotEmpty(pluginZKs)) {
@@ -237,6 +271,7 @@ public class PluginServiceImpl implements PluginService {
         return 0;
     }
 
+
     /**
      * sync plugin.
      *
@@ -250,7 +285,14 @@ public class PluginServiceImpl implements PluginService {
         zkClient.writeData(pluginPath, new PluginZkDTO(pluginDO.getId(),
                 pluginDO.getName(), pluginDO.getRole(), pluginDO.getEnabled()));
 
-        List<String> selectorZKs = zkClient.getChildren(ZkPathConstants.buildSelectorParentPath(pluginDO.getName()));
+
+        final String selectorParentPath = ZkPathConstants.buildSelectorParentPath(pluginDO.getName());
+
+        if (!zkClient.exists(selectorParentPath)) {
+            zkClient.createPersistent(selectorParentPath, true);
+        }
+
+        List<String> selectorZKs = zkClient.getChildren(selectorParentPath);
         selectorMapper.selectByQuery(new SelectorQuery(pluginDO.getId(), null)).forEach(selectorDO -> {
             if (CollectionUtils.isNotEmpty(selectorZKs)) {
                 selectorZKs.remove(selectorDO.getId());
@@ -266,7 +308,13 @@ public class PluginServiceImpl implements PluginService {
                     selectorDO.getName(), selectorDO.getMatchMode(), selectorDO.getType(), selectorDO.getSort(), selectorDO.getEnabled(),
                     selectorDO.getLoged(), selectorDO.getContinued(), selectorDO.getHandle(), selectorConditionZkDTOs));
 
-            List<String> ruleZKs = zkClient.getChildren(ZkPathConstants.buildRuleParentPath(pluginDO.getName()));
+
+            final String ruleParentPath = ZkPathConstants.buildRuleParentPath(pluginDO.getName());
+
+            if (!zkClient.exists(ruleParentPath)) {
+                zkClient.createPersistent(ruleParentPath, true);
+            }
+            List<String> ruleZKs = zkClient.getChildren(ruleParentPath);
             ruleMapper.selectByQuery(new RuleQuery(selectorDO.getId(), null)).forEach(ruleDO -> {
                 if (CollectionUtils.isNotEmpty(ruleZKs)) {
                     ruleZKs.remove(selectorDO.getId() + ZkPathConstants.SELECTOR_JOIN_RULE + ruleDO.getId());
